Move pytest fixtures into conftest.py
